Liquibase templates + pojo common WIP

+++ schema/src/main/java/art/limitium/sofa/ext/FBFactoryConverter.java
@@ -20,7 +20,7 @@ public class FBFactoryConverter implements TypeConverter {
             case Type.BytesType bytesType -> "CreateArray";
             case Type.DateType dateType -> "int";
             case Type.TimeMillisType timeMillisType -> "int";
-            case Type.TimeMicrosType timeMicrosType -> "int";
+            case Type.TimeMicrosType timeMicrosType -> "long";
             case Type.IntType intType -> "int";
             case Type.TimestampMillisType timestampMillisType -> "long";
             case Type.TimestampMicrosType timestampMicrosType -> "long";
