+++ schema/src/main/java/art/limitium/sofa/ext/FBTypeConverter.java
@@ -20,7 +20,7 @@ public class FBTypeConverter implements TypeConverter {
             case Type.BytesType bytesType -> "[byte]";
             case Type.DateType dateType -> "int";
             case Type.TimeMillisType timeMillisType -> "int";
-            case Type.TimeMicrosType timeMicrosType -> "int";
+            case Type.TimeMicrosType timeMicrosType -> "long";
             case Type.IntType intType -> "int";
             case Type.TimestampMillisType timestampMillisType -> "long";
             case Type.TimestampMicrosType timestampMicrosType -> "long";
